Update with rewrite.
Need program to convert both version 1 and 2 of the GPCP files.  Also
needed the CSV file in a specific format for input into a database.
This lead to the modularization of the code.  gpcp_parsers is a set of
classes for each type of conversion and should make creating more
converters easier. -- jms 19 Oct 2012